+++ test/testconfdb.c
@@ -60,11 +60,11 @@ static void print_config_tree(confdb_handle_t handle, hdb_handle_t parent_object
 {
 	hdb_handle_t object_handle;
 	char object_name[1024];
-	int object_name_len;
+	size_t object_name_len;
 	char key_name[1024];
-	int key_name_len;
+	size_t key_name_len;
 	char key_value[1024];
-	int key_value_len;
+	size_t key_value_len;
 	int res;
 	int i;
 
@@ -198,7 +198,7 @@ int main (int argc, char *argv[]) {
 	int result;
 	hdb_handle_t totem_handle;
 	char key_value[256];
-	int value_len;
+	size_t value_len;
 
 	result = confdb_initialize (&handle, &callbacks);
 	if (result != CS_OK) {
